requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.46.0
pydantic==2.7.4
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
python-dateutil==2.9.0.post0
icalendar==5.0.12
pytest==8.3.2
Jinja2==3.1.4
